Allow deep linking directly to a board.
Also refactored the API controller a tiny bit to use respond_with, which
mostly worked awesome.

It tried to build a URL for the created board, which is neat but not
quite correct by default in this case.  I give it the API path for the
board, though it's not used at all by ember.