Drop periodically failing timing-based test.
It would be awesome if this thing would work, but sometimes Ember gets
ahead of the Capybara tests where it checks on the link while the model
creation is still in flight.  Maybe you can replace this with a test on
the template itself.  Having the loading screen is important so that
nobody can start trying to manage the unpersisted board.