Can fetch and save configurable attributes.
The Ember part went really smooth.  I have to jump through a few hoops
when I represent a data structure like this as a resource, but it works.

Had to change the ConfiguredAttribute's id field from
"source_name.attribute_key" to "source_name:attribute_key" because the
dot was getting special treatment from Rails.  I needed a divider that
wouldn't exist in the name but would be passed straight through by
Rails.