Retool the board show page to reflect slide status.
This will serve as a place to tell people what to do next, or redirect
to the first slide is any exist.

Of course there seems to be a new bug with the way the slides
association in ember works when a board gets created while the app is
running.  I don't think it's my fault, either.  Meh.