Save when a pane is dragged to a new position.
And here's where things get tricky in testing land.  I'm not dragging
into a regular droptarget.  I'm being partly lazy, certainly, with not
making an attempt to unit test some of these view or controller methods.
But they're so tied together that I'm not sure how helpful it will be
without an integration test.