Enforce a width on a slide.
It was just kind of moving based on the browser window, which was
changing the number of possible columns.  So you wouldn't get a
consistent experience on a narrower browser.  That's nice for reading,
but if you're editing on a narrower browser you don't want to break it
for everybody else.

I think when you want to support mobile you need to make sure that
nobody can drag stuff around.