Allow deletion of a perspective assignment.
Not bad.  The main problem was converting the PerspectiveAssignmentView
to a component so that it could act in a controllerish fashion,
responding to the remove button.  I guess, now that I think about it, I
could have bubbled the action up to SlideController.  Then I could have
done the locking I wanted to do but was unable to while the delete was
in flight.  Meh.  Leaving it as-is for now.

I like the component, but one annoyance was the fact that it was
required to have a template to be registered with Ember.  The contents
of that template were previously dynamic, and I was going to use a
layout for the frame.  Now I have to render that 'anonymous' Em.View
inside the template for the PerspectiveAssignmentComponent.