Fix assignment removal bug.
A perspective assignment wasn't being removed from the interface
properly if it had been added in that running application instance.  It
would make the API call and delete it out of the database, but it would
continue to be displayed.  And then regridding and other stuff would be
way out of whack.