Update to Ember 1.1.2.  Some minor troubles.
Not really sure what it's up to here.  Suddenly it started creating new
records for each assigned perspective.  Not sure if there's a bug fixed
or added here, since this code is probably not perfect given the
difficulty of integrating gridster.

I did discover that it rerenders the view a lot more often than I
expected, which is lame.  I would like to come back and fix it.  I think
there are still some issues running around here where somebody can try
and reorder the grid while the assignment is still in flight.  I have a
bad feeling that the battle with this code is far from finished.