Jobs can write data associated with the assignment.
Bunch of changes that don't actually do much yet.  The job has access to
a data method that will be written to redis, associated with that
particular PerspectiveAssignment.  In a little bit I will serialize that
data and send it along in the serializer so that it can be displayed.

Also refactored (sort of) the PerspectiveAssignment class.  It still has
all the same responsibilities, but the different aspects are moved into
concerns.