Show real data in the templates.
Ouch.  First, what is called "data" in the controllers was switched to
"payload" midstream because it seems like it's a reserved ember term.
So that was fun to sort out.

The databinding stuff wasn't that bad.  The payloads are real model
objects in Ember so that they can be easily updated with new data later
and have the view reflect the changes.

I figured out that I can sideload the payload in the create action to
stop if from making a request, which I should probably tweak the
configurable attributes to do as well.  I still might need the API
endpoint for deep linking on that one, though.