Move perspective configuration to a reasonable spot.
It was just jammed above the template it was controlling.  Fun for
debugging, no good for displaying.  Moved it above the grid, which is
going to result is some screwy UI where we don't want to let you see the
delete/configure controls for that particular assignment while you're
already configuring it.  Later.