Fix assignments acting flakey after create.
Well that was a really fun bug.  Since it's completely undocumented, you
have to guess, and it jukes you by working 95% of the way with one
naming convention and 100% the other way.  The assignment was getting
persisted properly, but wasn't picking up its new ID after the AJAX call
came back.  Wow.