Sideload assignment configuration with perspectives.
It keeps it from making a second request to get the configuration for
the assignment that it just created.  That's nice and all, but I was
actually just trying to troubleshoot this godawful issue where new
assignments are causing JS errors for every data update thereafter.
It's when it's rendering the template, but other than that I can't
figure it out.  The stacktrace is absolutely useless.