Stop fighting Bootstrap's less support.
If I just wanted to tweak some colors or load a subset of the styles, it
would work.  But the point was to try and use fewer bootstrap-specific
styles in my markup, and it didn't do much there.  Some components can
be less'd up, some can't.  I'd rather have fewer moving parts and some
bad markup than extra gems and frameworks just to clean up the few bits
of markup that can work with less.