Don't strip extension from asset paths.
You don't have to.  It works the same whether you link to
"myfile.css.scss" or "myfile".  Not doing the flakey splitting on "."
will probably cause fewer crazy bugs.  I'm particularly thinking about
how, because of the way the assets works, I think that a stylesheet
named "base" in one source will overwrite a "base" in another source.
I'm going to suggest in the docs to namespace your stylesheets, and some
people might use a "." to do it.